Refactor Elastic initialisation in BE
* Initialise DB pool once, reuse across modules
* Ensure all connections to Elastic are consistently asynchronous
